requirements.txt
PyQt5~=5.15.6
numpy~=1.21.2
Pillow~=8.4.0
pytest~=7.1.3